edit the structure of project by changing the name of packages to standard naming convention + edit some extra includes and remove them all from header files + change the commenting of project to the doxygen standard and make the whole project well commented!
