+++ CMakeLists.txt
@@ -3,4 +3,4 @@ project(cplusclass1399)
 
 set(CMAKE_CXX_STANDARD 14)
 
-add_executable(cplusclass1399 main.cpp src/FileSender_Package/FileSender.h src/FileSender_Package/Compression_Package/Compression.h src/FileSender_Package/Encryption_Package/Encryption.h test/UnitTest.cpp test/UnitTest.h)
+add_executable(cplusclass1399 main.cpp src/file-sender/FileSender.h src/file-sender/compression/Compression.h src/file-sender/encryption/Encryption.h test/UnitTest.cpp test/UnitTest.h)
